Light refactor to ease merge from master
