+++ core/auth/index.ts
@@ -7,8 +7,9 @@ import { z } from 'zod';
 import { client } from '~/client';
 import { graphql, ResultOf } from '~/client/graphql';
 import { getCartId } from '~/lib/cart';
+import { loginWithB2B } from './loginWithB2B';
 
-const LoginMutation = graphql(`
+export const LoginMutation = graphql(`
   mutation LoginMutation($email: String!, $password: String!, $cartEntityId: String) {
     login(email: $email, password: $password, guestCartEntityId: $cartEntityId) {
       customerAccessToken {
@@ -86,16 +87,16 @@ async function loginWithPassword(
     return null;
   }
 
+  const b2bToken = await loginWithB2B({
+    customer: result.customer,
+    customerAccessToken: result.customerAccessToken,
+  });
+
   return {
     name: `${result.customer.firstName} ${result.customer.lastName}`,
     email: result.customer.email,
     customerAccessToken: result.customerAccessToken.value,
-    ...(process.env.B2B_API_TOKEN && {
-      b2bToken: await loginWithB2B({
-        customer: result.customer,
-        customerAccessToken: result.customerAccessToken,
-      }),
-    }),
+    b2bToken,
   };
 }
 
@@ -122,63 +123,18 @@ async function loginWithJwt(jwt: string, cartEntityId?: string): Promise<User | 
     return null;
   }
 
+  const b2bToken = await loginWithB2B({
+    customer: result.customer,
+    customerAccessToken: result.customerAccessToken,
+  });
+
   return {
     name: `${result.customer.firstName} ${result.customer.lastName}`,
     email: result.customer.email,
     customerAccessToken: result.customerAccessToken.value,
     impersonatorId,
-    ...(process.env.B2B_API_TOKEN && {
-      b2bToken: await loginWithB2B({
-        customer: result.customer,
-        customerAccessToken: result.customerAccessToken,
-      }),
-    }),
+    b2bToken,
   };
-}
-
-interface LoginWithB2BParams {
-  customer: NonNullable<ResultOf<typeof LoginMutation>['login']['customer']>;
-  customerAccessToken: NonNullable<ResultOf<typeof LoginMutation>['login']['customerAccessToken']>;
-}
-
-async function loginWithB2B({ customer, customerAccessToken }: LoginWithB2BParams) {
-  if (!process.env.B2B_API_TOKEN) {
-    throw new Error('Environment variable B2B_API_TOKEN is not set');
-  }
-
-  const channelId = process.env.BIGCOMMERCE_CHANNEL_ID;
-
-  const payload = {
-    channelId,
-    customerId: customer.entityId,
-    customerAccessToken,
-  };
-
-  const response = await fetch(`https://api-b2b.bigcommerce.com/api/io/auth/customers/storefront`, {
-    method: 'POST',
-    headers: {
-      Accept: 'application/json',
-      'Content-Type': 'application/json',
-      authToken: process.env.B2B_API_TOKEN,
-    },
-    body: JSON.stringify(payload),
-  });
-
-  const B2BTokenResponseSchema = z.object({
-    data: z.object({
-      token: z.array(z.string()),
-    }),
-  });
-
-  const {
-    data: { token },
-  } = B2BTokenResponseSchema.parse(await response.json());
-
-  if (!token[0]) {
-    throw new Error('No token returned from B2B API');
-  }
-
-  return token[0];
 }
 
 async function authorize(credentials: unknown): Promise<User | null> {
@@ -214,15 +170,11 @@ const config = {
     jwt: ({ token, user }) => {
       // user can actually be undefined
       // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-      if (!user) {
-        return token;
-      }
-
-      if (user.customerAccessToken) {
+      if (user?.customerAccessToken) {
         token.customerAccessToken = user.customerAccessToken;
       }
 
-      if (user.b2bToken) {
+      if (user?.b2bToken) {
         token.b2bToken = user.b2bToken;
       }
 
@@ -243,7 +195,6 @@ const config = {
   events: {
     async signOut(message) {
       const customerAccessToken = 'token' in message ? message.token?.customerAccessToken : null;
-      // @todo check if b2bToken is also valid?
 
       if (customerAccessToken) {
         try {
